TASK-1: Add CRUD functionality for projects
I choose to unify the AddProject.vue and EditProject.vue into UpsertProject.vue, since the code duplicated was simply too much. If the difference between the 2 different use cases changes, it would be trivial to create sub-components inside the UpsertProject.vue file to handle any differences.
I also added server-side validation for the creation and updating of the project.